+++ src/json/ndjson/errors.rs
@@ -5,10 +5,10 @@ pub mod collection;
 /// Wrapper around the various errors we can encounter while processing the data
 #[derive(Error, Debug)]
 pub enum NDJSONError {
-    #[error("failed to read input data")]
+    #[error("Failed to read input data")]
     IOError(#[from] std::io::Error),
-    #[error("line failed to parse as valid JSON")]
+    #[error("Line failed to parse as valid JSON")]
     JSONParsingError(#[from] serde_json::Error),
-    #[error("line returned empty for ther given query")]
+    #[error("Line returned empty for ther given query")]
     EmptyQuery,
 }