+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - Switched to `serde_json_path` for JSON path querying
+- Added anyhow for error handling
 
 ## [0.5.5] - 2023-07-12